Archiving project
The project is currently not being actively developed anymore. While libtrashcan works, there are implementation and design choices that I am not particularly happy with, but currently do not have the time to address. Since nearly no one uses the library and thus developers do not rely on its maintenance, the project is archived for now.